@@ -1,47 +1,42 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <ApplicationIcon />
-    <OutputType>WinExe</OutputType>
-    <StartupObject />
-    <AssemblyName>ElasticErrorRates.API</AssemblyName>
-    <RootNamespace>ElasticErrorRates.API</RootNamespace>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<AssemblyName>ElasticErrorRates.API</AssemblyName>
+		<RootNamespace>ElasticErrorRates.API</RootNamespace>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <Content Remove="appsettings.json" />
-    <Content Remove="appsettings.Preprod.json" />
-  </ItemGroup>
+	<ItemGroup>
+		<Content Remove="appsettings.json" />
+		<Content Remove="appsettings.Preprod.json" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="wwwroot\" />
-  </ItemGroup>
+	<ItemGroup>
+		<Folder Include="wwwroot\" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.3" />
-	<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.4" />
-	<PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.13" />
+		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="7.0.11" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.13" />
+		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<EmbeddedResource Include="appsettings.json">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</EmbeddedResource>
+		<EmbeddedResource Include="appsettings.Preprod.json">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</EmbeddedResource>
+	</ItemGroup>
 
-  <ItemGroup>
-    <EmbeddedResource Include="appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </EmbeddedResource>
-    <EmbeddedResource Include="appsettings.Preprod.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </EmbeddedResource>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ElasticErrorRates.Core\ElasticErrorRates.Core.csproj" />
-    <ProjectReference Include="..\ElasticErrorRates.Hangfire\ElasticErrorRates.Hangfire.csproj" />
-    <ProjectReference Include="..\ElasticErrorRates.Injection\ElasticErrorRates.Injection.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\ElasticErrorRates.Core\ElasticErrorRates.Core.csproj" />
+		<ProjectReference Include="..\ElasticErrorRates.Hangfire\ElasticErrorRates.Hangfire.csproj" />
+		<ProjectReference Include="..\ElasticErrorRates.Injection\ElasticErrorRates.Injection.csproj" />
+	</ItemGroup>
 
 </Project>
