@@ -1,5 +1,4 @@
-﻿using System;
-using ElasticErrorRates.Core.CQRS.Command;
+﻿using ElasticErrorRates.Core.CQRS.Command;
 
 namespace ElasticErrorRates.CQRS.Command
 {
