@@ -1,7 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace ElasticErrorRates.Core.CQRS.Command
+﻿namespace ElasticErrorRates.Core.CQRS.Command
 {
     // <summary>
     /// Base interface for command handlers
