@@ -1,7 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace ElasticErrorRates.Core.CQRS.Query
+﻿namespace ElasticErrorRates.Core.CQRS.Query
 {
     public interface IQueryDispatcher
     {
