@@ -1,9 +1,6 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace ElasticErrorRates.Core.CQRS.Query
+﻿namespace ElasticErrorRates.Core.CQRS.Query
 {
-   
+
     public interface IQueryHandler
     {
         Task<TResult> RetrieveAsync<TResult>(Func<Task<TResult>> query);
