@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-
-namespace ElasticErrorRates.Core.CQRS.Query
+﻿namespace ElasticErrorRates.Core.CQRS.Query
 {
     public interface IQueryResult<T> where T : class
     {
