@@ -1,7 +1,4 @@
-﻿using System;
-using ElasticErrorRates.Core.Enums;
-
-namespace ElasticErrorRates.Core.Criteria.Dashboard
+﻿namespace ElasticErrorRates.Core.Criteria.Dashboard
 {
     public class LogQuantityCriteria : LogCriteria
     {
