@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using ElasticErrorRates.Core.Models;
+﻿using ElasticErrorRates.Core.Models;
 using Nest;
 
 namespace ElasticErrorRates.Core.Persistence
