@@ -1,6 +1,4 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using ElasticErrorRates.Core.Models;
+﻿using ElasticErrorRates.Core.Models;
 using Microsoft.EntityFrameworkCore;
 
 namespace ElasticErrorRates.Persistence.Context
